refactor: move Hero to own component folder, add hero image assets
